move logic for GrypeDBSyncTask to GrypeDBSyncManager and incorporate us in feeds.py
